requirements.txt
pygrametl>=2.8.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
